+++ src/main/java/com/caiza/clinical_alerts/telemetry/rules/HateRateRulesStrategy.java
@@ -2,6 +2,8 @@ package com.caiza.clinical_alerts.telemetry.rules;
 
 import com.caiza.clinical_alerts.telemetry.event.TelemetryReceivedEvent;
 
+import java.math.BigDecimal;
+
 public class HateRateRulesStrategy implements TelemetryRulesStrategy {
 
     @Override
@@ -11,12 +13,12 @@ public class HateRateRulesStrategy implements TelemetryRulesStrategy {
 
     @Override
     public RiskLevel evaluate(TelemetryReceivedEvent event) {
-        double value = event.measuredValue();
+        BigDecimal value = event.measuredValue();
 
-        if (value < 40 || value > 140) {
+        if (value.intValue() < 40 || value.intValue() > 140) {
             return RiskLevel.CRITICAL;
         }
-        if (value < 50 || value > 120){
+        if (value.intValue() < 50 || value.intValue() > 120){
             return RiskLevel.WARNING;
         }
         return RiskLevel.NORMAL;
