+++ src/main/java/com/caiza/clinical_alerts/telemetry/rules/OxygenRateRulesStrategy.java
@@ -2,6 +2,8 @@ package com.caiza.clinical_alerts.telemetry.rules;
 
 import com.caiza.clinical_alerts.telemetry.event.TelemetryReceivedEvent;
 
+import java.math.BigDecimal;
+
 public class OxygenRateRulesStrategy implements TelemetryRulesStrategy {
 
     @Override
@@ -11,12 +13,12 @@ public class OxygenRateRulesStrategy implements TelemetryRulesStrategy {
 
     @Override
     public RiskLevel evaluate(TelemetryReceivedEvent event) {
-        double value = event.measuredValue();
+        BigDecimal value = event.measuredValue();
 
-        if (value < 85) {
+        if (value.intValue() < 85) {
             return RiskLevel.CRITICAL;
         }
-        if (value > 92){
+        if (value.intValue() > 92){
             return RiskLevel.WARNING;
         }
         return RiskLevel.NORMAL;
